Se ha vuelto a cambiar el atributo del tamaño de los ficheros y del parametro del espacio donado, en vez de enteros, trabajamos con longs. Tambien se han implementado y testeado las funciones que procesan los mensajes de petición de un fichero (tipo 0) y los mensajes para almacenar un fichero (tipo 1), utilizando los servicios ya disponibles del servidor.

+++ src/main/java/es/udc/fic/tic/nautilus/config/ConfigHandler.java
@@ -22,7 +22,7 @@ public class ConfigHandler {
 		File fileConfig = new File("config.xml");
 		
 		Boolean serverAvaliable = null;
-		int limitSpace = 0;
+		long limitSpace = 0;
 		ModelConstanst.LANGUAGE language = null;
 		List<ServerInfo> serverPreferences = new ArrayList<ServerInfo>();
 		String homeFolder = null;
@@ -40,7 +40,7 @@ public class ConfigHandler {
 			if (config.getChildTextTrim("limitSpace").equals("0")) {
 				limitSpace = 0;
 			} else {
-				limitSpace = Integer.parseInt(config.getChildTextTrim("limitSpace"));
+				limitSpace = Long.parseLong(config.getChildTextTrim("limitSpace"));
 			}
 			
 			if (config.getChildTextTrim("language").equals("EN")) {
